requirements.txt
streamlit>=1.44,<2
requests>=2.31,<3
pyserial>=3.5,<4
